pyproject.toml
[project]
name = "agno-data-science-team"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=2.3.21",
    "ddgs>=9.10.0",
    "duckdb>=1.4.3",
    "fastapi[standard]>=0.128.0",
    "google-genai>=1.56.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.0",
    "openai>=2.14.0",
    "pandas>=2.3.3",
    "pathlib>=1.0.1",
    "pip>=25.3",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
